Loot UI Update + Overworld camera follow
Loot UI now has spots for showing the new (and old) limb names and effects.
Currently it's just text that's manually written there. None of the text actually reflects what the player has equipped. The loot is not generated based on the enemy either.

Camera now follows the player in the overworld.