"Playable" build
Changed some code to make it work when built.

Also fixed an annoying bug which was prevent the checkpoint system from getting the correct value from the singleton sometimes...? Still not sure what the cause is, but I fixed by getting the singleton via the Gameobject's name rather than the component type.